Add container to serve the built Nuxt site
This container serves the same assets as the production site. We have
found there are some differences between the development Nuxt app and
the built production site. This will allow better local testing.
