+++ README.md
@@ -21,14 +21,19 @@ Install app dependencies:
 ```
 docker compose run ruby-script bundle install
 docker compose run script yarn
-docker compose run script bash -c "cd server && yarn"
+docker compose run script bash -c "cd preview-nuxt && yarn"
 ```
 ## Local Docker Usage
 
-Run the servers:
+For local development (with hot reloading):
 
 ```
 docker compose up
 ```
 
-Preview the site at http://0.0.0.0:3001.
+Preview the Jekyll site at http://0.0.0.0:4000.
+Preview the Nuxt site at http://0.0.0.0:3000.
+
+In production we build and serve static files which is different to the Nuxt development server.
+You can preview the static Nuxt site at http://0.0.0.0:5001.
+Note this does not support hot reloading.
