+++ docker-compose.yml
@@ -5,20 +5,25 @@ services:
     image: node:16
     working_dir: /work
     volumes: [ ".:/work" ]
-    command: "yarn start"
+    command:
+      - bash
+      - -c
+      - yarn generate && yarn start
     ports: [ "3000:3000", "24678:24678" ]
+    environment:
+      - "ENV=development"
   legacy:
     image: ruby:2.7
     working_dir: /work/legacy
     volumes: [ ".:/work" ]
-    command: "bundle exec jekyll serve --livereload --host=0.0.0.0 --port 3001"
-    ports: [ "4000:3001", "35729:35729" ]
-  server:
+    command: "bundle exec jekyll serve --livereload --host=0.0.0.0"
+    ports: [ "4000:4000", "35729:35729" ]
+  preview-nuxt:
     image: node:16
-    working_dir: /work/server
+    working_dir: /work/preview-nuxt
     volumes: [ ".:/work" ]
     command: "yarn start"
-    ports: [ "3001:3002" ]
+    ports: [ "5001:5001" ]
   script:
     image: node:16
     working_dir: /work
